requirements.txt
streamlit==1.33.0
torch==2.3.0
torchvision==0.18.0
matplotlib==3.8.4
Pillow==10.3.0